# api/swag/swagger.yaml
definitions:
  converter.ConvertByGavingKeyRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/converter.ConvertByGavingKeyRequestItem'
        maxItems: 10
        minItems: 1
        type: array
    required:
    - items
    type: object
  converter.ConvertByGavingKeyRequestItem:
    properties:
      sourceKey:
        description: '@description 如果这个值为合法url就会根据url地址获取源文件，如果不是则默认为文件key，通过minio获取文件'
        type: string
      targetFileDpi:
        type: integer
      targetKey:
        type: string
    required:
    - sourceKey
    - targetKey
    type: object
  converter.ConvertByGavingKeyResponse:
    properties:
      isAllSucceed:
        type: boolean
      items:
        items:
          $ref: '#/definitions/converter.ConvertByGavingKeyResponseItem'
        type: array
    required:
    - isAllSucceed
    - items
    type: object
  converter.ConvertByGavingKeyResponseItem:
    properties:
      isSucceed:
        type: boolean
      message:
        type: string
      sourceKey:
        type: string
      targetFileSize:
        type: integer
    required:
    - isSucceed
    - sourceKey
    type: object
  imginfo.GetImgInfoItemResponse:
    properties:
      fileKey:
        type: string
      isSucceed:
        type: boolean
      message:
        type: string
      width:
        type: integer
    required:
    - fileKey
    type: object
  imginfo.GetImgInfoRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/imginfo.GetImgInfoRequestItem'
        maxItems: 20
        minItems: 1
        type: array
    required:
    - items
    type: object
  imginfo.GetImgInfoRequestItem:
    properties:
      fileKey:
        description: '@description minio文件key'
        type: string
    type: object
  imginfo.GetImgInfoResponse:
    properties:
      isAllSucceed:
        type: boolean
      items:
        items:
          $ref: '#/definitions/imginfo.GetImgInfoItemResponse'
        maxItems: 20
        minItems: 1
        type: array
    required:
    - items
    type: object
  models.AclError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
      requestStack:
        type: string
      validationErrors:
        items:
          $ref: '#/definitions/models.ValidationError'
        type: array
    type: object
  models.CommandResponse:
    properties:
      isSuccessd:
        type: boolean
      message:
        type: string
    type: object
  models.CommonErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.AclError'
    type: object
  models.ValidationError:
    properties:
      members:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  tar.PackRequest:
    properties:
      fileKey:
        type: string
      isGziped:
        type: boolean
      items:
        items:
          $ref: '#/definitions/tar.PackRequestItem'
        type: array
    type: object
  tar.PackRequestItem:
    properties:
      fileKey:
        description: '@description 如果此参数是一个合法的url则会根据url获取文件，否则视为minio文件key'
        type: string
      fileName:
        description: 文件夹层级用"/"表示
        type: string
      lastModifyTime:
        description: ISO 8601 时间
        type: string
    type: object
info:
  contact: {}
paths:
  /Converter/GetFisrtImageByGavingKey:
    post:
      consumes:
      - application/json
      description: 根据文件key或者文件url获取文件并转成相应的目标图片，只转第一个图层或者第一页，目前支持psd->jpeg;ai->jpeg;pdf->jpeg
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/converter.ConvertByGavingKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/converter.ConvertByGavingKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: GetFisrtImageByGavingKey
      tags:
      - GetFisrtImageByGavingKey
  /Converter/GetPdfByGavingKey:
    post:
      consumes:
      - application/json
      description: 根据文件key或者文件url获取文件并转成相应的目标pdf，只转第一个图层或者第一页，目前支持psd->pdf;ai->pdf;pdf->pdf
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/converter.ConvertByGavingKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/converter.ConvertByGavingKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: GetPdfByGavingKey
      tags:
      - GetPdfByGavingKey
  /GetImgInfo:
    post:
      consumes:
      - application/json
      description: 批量获取图片宽高，只支持png,jpg,gif类型图片
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/imginfo.GetImgInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/imginfo.GetImgInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: GetImgInfo
      tags:
      - GetImgInfo
  /Tar/Pack:
    post:
      consumes:
      - application/json
      description: 根据文件key或者文件url归档，或者归档压缩生成文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tar.PackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: PackByGavingKey
      tags:
      - PackByGavingKey
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"